% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_ctrl_cmp_OR.R
\name{clu_ctrl_compare}
\alias{clu_ctrl_compare}
\title{Cluster Vs Control odds ratio for binary outcome using logistic regression, optionally control for covariates.}
\usage{
clu_ctrl_compare(
  data,
  cluster_colname,
  ctrl_label,
  outcome_colname,
  covariate_colname = c()
)
}
\arguments{
\item{data}{a dataframe containing cluster, outcome, and optionally covariates.}

\item{cluster_colname}{name of a column containing clustering information. Number of unique clusters should be at least 2.}

\item{ctrl_label}{label for control group. Should be one of the labels in cluster column.}

\item{outcome_colname}{name of a column containing binary outcome.}

\item{covariate_colname}{a vector of names of columns for covariates. Default is empty.}
}
\value{
a dataframe containing enrichment information
}
\description{
Cluster Vs Control odds ratio for binary outcome using logistic regression, optionally control for covariates.
}
