% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_EL.R
\name{deprecated_utility_EL_advanced}
\alias{deprecated_utility_EL_advanced}
\title{Plot network from data file using ExplodeLayout with multiple graphical parameters and save the plot to pdf}
\usage{
deprecated_utility_EL_advanced(
  inputName_nodelist = "example_nodelist.csv",
  inputName_incidmat = "example_incidmat.csv",
  radius = 1.5,
  outputName_plotpdf = "example_plot.pdf",
  plotlabel = "col",
  nodesize_min = 1,
  nodesize_max = 10,
  nodesize_by_degree = "row",
  nodeborder_row = 0.1,
  nodeborder_col = 2,
  nodelabelsize = 3,
  edgethickness_min = 0.01,
  edgethickness_max = 1,
  edgegrayscale_min = 0.5,
  edgegrayscale_max = 1,
  edgethickness_by_weight = TRUE
)
}
\arguments{
\item{inputName_nodelist}{name of nodelist file}

\item{inputName_incidmat}{name of incidence matrix file}

\item{radius}{radius of ExplodeLayout. Default is 1.5}

\item{outputName_plotpdf}{name of the pdf to plot.}

\item{plotlabel}{whether to plot node labels. 'neither' labels no nodes, 'row' labels row nodes, 'col' labels column nodes, 'both' labels all nodes. Default is 'col'.}

\item{nodesize_min}{minimum size of node. Also used if not plot node size by degree.}

\item{nodesize_max}{minimum size of node.}

\item{nodesize_by_degree}{whether to plot node size by degree. Can be 'neither', 'row', 'col', or 'both'. Default is 'row'.}

\item{nodeborder_row}{stroke of row node. Default is 1.}

\item{nodeborder_col}{stroke of column node. Default is 2.}

\item{nodelabelsize}{size of label. Default is 3.}

\item{edgethickness_min}{minimum thickness of an edge. Also used if edge thickness not by weight. Default is 0.01.}

\item{edgethickness_max}{maximum thickness of an edge.}

\item{edgegrayscale_min}{darkest color of an edge on grey scale. The smaller the darker. Default is 0.5.}

\item{edgegrayscale_max}{lightest color of an edge on grey scale. Default is 1.}

\item{edgethickness_by_weight}{whether to plot edge width by thickness.}
}
\value{
the ggplot object

This function does the following things:
1. reads nodelist file from [inputName_nodelist];
2. reads incidence matrix file from [inputName_incidmat];
3. visualizes the network using ExplodeLayout according to user specified parameters and prints the plot on screen;
4. saves the plot to [outputName_plotpdf].
}
\description{
Plot network from data file using ExplodeLayout with multiple graphical parameters and save the plot to pdf
}
