% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_ctrl_cmp_OR.R
\name{enrich_CluVsCtrl}
\alias{enrich_CluVsCtrl}
\title{Do cluster Vs Control enrichment for a binary outcome, optionally controlling for covariates.}
\usage{
enrich_CluVsCtrl(
  data,
  cluster_colname,
  ctrl_label,
  outcome_colname,
  covariate_colname = c(),
  weight_colname = NA
)
}
\arguments{
\item{data}{a dataframe containing cluster, outcome, and optionally covariates.}

\item{cluster_colname}{name of a column containing clustering information. Number of unique clusters should be at least 2.}

\item{ctrl_label}{label for control group. Should be one of the labels in cluster column.}

\item{outcome_colname}{name of a column containing binary outcome.}

\item{covariate_colname}{a vector of names of columns for covariates. Default is empty.}

\item{weight_colname}{weight for regression, optional}
}
\value{
a list of 3 dataframes. 'ccc' is the full enrichment result. 'ccc_sig' only shows significant ones after Bonferroni correction. 'tb' is a simple counting table with proportions of outcome for each cluster.
}
\description{
Do cluster Vs Control enrichment for a binary outcome, optionally controlling for covariates.
}
