% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection_UniVar_LogReg.R
\name{FeatureSelection_UniVar_LogReg}
\alias{FeatureSelection_UniVar_LogReg}
\title{Univariable feature selection using logistic regression.}
\usage{
FeatureSelection_UniVar_LogReg(
  data,
  independent_variables,
  covariates = c(),
  outcome
)
}
\arguments{
\item{data}{dataframe containing independent variables and outcome for feature selection}

\item{independent_variables}{names of independent variables}

\item{covariates}{optional}

\item{outcome}{name of outcome}
}
\value{
dataframe containing univariable feature selection results in details
}
\description{
Univariable feature selection using logistic regression.
}
