% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection_UniVar_LogReg.R
\name{FeatureSelection_UniVar_LogReg_MultiOutcomes}
\alias{FeatureSelection_UniVar_LogReg_MultiOutcomes}
\title{Summarize univariable feature selection results (Bonferroni corrected p-values) for multiple outcomes.}
\usage{
FeatureSelection_UniVar_LogReg_MultiOutcomes(
  data,
  independent_variables,
  covariates = c(),
  outcomes
)
}
\arguments{
\item{data}{dataframe containing independent variables and outcome for feature selection}

\item{independent_variables}{names of independent variables}

\item{covariates}{optional, same for all independent variables.}

\item{outcomes}{names of outcomes}
}
\value{
dataframe containing Bonferroni corrected p-values from univariable feature selection for each outcome.
}
\description{
Summarize univariable feature selection results (Bonferroni corrected p-values) for multiple outcomes.
}
