% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cluster_cmp_OR.R
\name{flat_2x2_01table}
\alias{flat_2x2_01table}
\title{Generate a 2x2 table for two binary variables in flat form}
\usage{
flat_2x2_01table(x, y)
}
\arguments{
\item{x}{a binary variable}

\item{y}{a binary variable of same length}
}
\value{
a vector of 4 numbers counting x==0&y==0, x==0&y==1, x==1&y==0, x==1&y==1
}
\description{
Generate a 2x2 table for two binary variables in flat form
}
