% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_edge_color}
\alias{get_edge_color}
\title{Get edge color (gray scale) from edge weight.}
\usage{
get_edge_color(edgeweight, mincolor = 0.5, maxcolor = 1)
}
\arguments{
\item{edgeweight}{a vector of values used to get edge width. The larger the edge weight, the darker the edge color.}

\item{mincolor}{darkest gray. 0 for black, 1 for white.}

\item{maxcolor}{brightest gray. 0 for black, 1 for white.}
}
\value{
a vector of edge color (gray scale)
}
\description{
Get edge color (gray scale) from edge weight.
}
