% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_edge_width}
\alias{get_edge_width}
\title{Calculate edge width from edge weight}
\usage{
get_edge_width(edgeweight, weighted = T, minwidth = 0.01, maxwidth = 1)
}
\arguments{
\item{edgeweight}{a vector of values used to get edge width. The larger the edge weight, the larger the edge width. The transformation from edge weight to edge width is linear.}

\item{weighted}{if true (default), scale edge width between minwidth and maxwidth. If false, use minwidth for all edges.}

\item{minwidth}{Default is 0.01}

\item{maxwidth}{Default is 1}
}
\value{
a vector of edge width
}
\description{
Calculate edge width from edge weight
}
