% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_edgelist_for_plotting}
\alias{get_edgelist_for_plotting}
\title{Get edgelist with extra features and plotting information.}
\usage{
get_edgelist_for_plotting(
  nodelist,
  incidence_matrix,
  edgethickness_min = 0.01,
  edgethickness_max = 1,
  edgecolor_min = 0.5,
  edgecolor_max = 1,
  edgethickness_by_weight = TRUE
)
}
\arguments{
\item{nodelist}{a data frame containing 'Label','X','Y'.}

\item{incidence_matrix}{with row names and column names.}

\item{edgethickness_min}{Also used if edgethickness_by_weight==FALSE. Default is 0.01.}

\item{edgethickness_max}{Default is 1}

\item{edgecolor_min}{Darkest in gray scale. 0 for black, 1 for white. Default is 0.5}

\item{edgecolor_max}{Brightest in gray scale. Default is 1.}

\item{edgethickness_by_weight}{if TRUE, scale edge width by edge value. Default is TRUE.}
}
\value{
a data frame containing 'nodesR',	'nodesC',	'values',	'x0',	'y0',	'x1',	'y1',	'weight',	'edgewidth',	'color'.
}
\description{
Get edgelist with extra features and plotting information.
}
