% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_edgelist_from_incidmat}
\alias{get_edgelist_from_incidmat}
\title{Get edgelist from incidence matrix. Identical to the function with same name in ExplodeLayout. Duplicated for user's convenience if user doesn't want to load ExplodeLayout.}
\usage{
get_edgelist_from_incidmat(incidence_matrix)
}
\arguments{
\item{incidence_matrix}{incidence matrix of a bipartite network. Should be a data frame with row names and column names.}
}
\value{
a data frame with 3 columns 'nodesR', 'nodesC', 'values', where each row contains the row node, column node, and value in incidence matrix. Only non-zero values are kept in the edgelist.
}
\description{
Get edgelist from incidence matrix. Identical to the function with same name in ExplodeLayout. Duplicated for user's convenience if user doesn't want to load ExplodeLayout.
}
