% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_node_color}
\alias{get_node_color}
\title{Get node color based on partition}
\usage{
get_node_color(partition, type = c("Cluster", "Entity", "Outcome"))
}
\arguments{
\item{partition}{a vector of integer used to color nodes}

\item{type}{can be 'Cluster', 'Entity', 'Outcome', each with predefined color code}
}
\value{
a vector of node colors
}
\description{
Get node color based on partition
}
