% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_node_degree}
\alias{get_node_degree}
\title{Get node degree from incidence matrix}
\usage{
get_node_degree(incidence_matrix)
}
\arguments{
\item{incidence_matrix}{a representation of bipartite network}
}
\value{
a vector representing degree of each node, ordered first by rows from top to bottom, then by columns from left to right.
}
\description{
Get node degree from incidence matrix
}
