% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_node_nodeborder}
\alias{get_node_nodeborder}
\title{Get node border thickness based on Entity}
\usage{
get_node_nodeborder(Entity, nodeborder_row = 1, nodeborder_col = 2)
}
\arguments{
\item{Entity}{indicating whether a node in bipartite network is on the row side (1) or column side (2).}

\item{nodeborder_row}{Default is 1}

\item{nodeborder_col}{Default is 2}
}
\value{
a vector of node border thickness used in ggplot2 as stroke.
}
\description{
Get node border thickness based on Entity
}
