% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_node_plotlabel}
\alias{get_node_plotlabel}
\title{Get node label to plot based on partition}
\usage{
get_node_plotlabel(Label, Entity, by = c("col", "row", "both", "neither"))
}
\arguments{
\item{Label}{original node labels}

\item{Entity}{indicating whether a node in bipartite network is on the row side (1) or column side (2).}

\item{by}{which side to keep. Can be 'col', 'row', 'both', 'neither'}
}
\value{
a vector of node labels where NA means don't plot.
}
\description{
Get node label to plot based on partition
}
