% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_node_shape}
\alias{get_node_shape}
\title{Get node shape based on Entity}
\usage{
get_node_shape(Entity, rowshape = 21, colshape = 24)
}
\arguments{
\item{Entity}{indicating whether a node in bipartite network is on the row side (1) or column side (2).}

\item{rowshape}{Default is 21 (filled circle)}

\item{colshape}{Default is 24 (filled triangle point-up)}
}
\value{
a vector of node shape encoded in base R (pch).
}
\description{
Get node shape based on Entity
}
\references{
http://www.sthda.com/english/wiki/r-plot-pch-symbols-the-different-point-shapes-available-in-r
}
