% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_node_size_by_degree}
\alias{get_node_size_by_degree}
\title{Get node size by degree.}
\usage{
get_node_size_by_degree(
  Degree,
  Entity,
  by = c("row", "col", "both", "neither"),
  minsize = 0,
  maxsize = 10
)
}
\arguments{
\item{Degree}{a numeric vector}

\item{Entity}{indicating whether a node in bipartite network is on the row side (1) or column side (2).}

\item{by}{Scale the degree by 'row', 'col', 'both', or 'neither'. Default is 'row'}

\item{minsize}{minimum size of a node. Also used if node size is not scaled by degree.}

\item{maxsize}{maximum size of a node.}
}
\value{
a vector of node size.
}
\description{
Get node size by degree.
}
