% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{get_nodelist}
\alias{get_nodelist}
\title{Get nodelist of a bipartite network with 2D coordinates and predefined cluster or partition.}
\usage{
get_nodelist(incid_mat, Cluster)
}
\arguments{
\item{incid_mat}{incidence matrix of the bipartite network.}

\item{Cluster}{a vector of positive integers indicating cluster membership of each node. Here is the order: first row nodes from top to bottom, then column nodes from left to right.}
}
\value{
a dataframe with 5 columns: Label, Cluster, X, Y, Entity. Label is the node name. X and Y are 2D coordinates from FR layout. Entity indicates whether a node is a row node (1) or a column node (2).
}
\description{
Get nodelist of a bipartite network with 2D coordinates and predefined cluster or partition.
}
