% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{get_nodelist_for_plotting}
\alias{get_nodelist_for_plotting}
\title{Get nodelist with extra features and plotting information. It overwrites the function with same name in ExplodeLayout for more detailed control.}
\usage{
get_nodelist_for_plotting(
  nodelist,
  new_coordinates,
  incidence_matrix,
  plotlabel = "col",
  nodesize_min = 1,
  nodesize_max = 10,
  nodesize_by_degree = "row",
  nodeborder_row = 0.1,
  nodeborder_col = 2,
  nodelabelsize = 3,
  nodelabelcolor = "grey20"
)
}
\arguments{
\item{nodelist}{a data frame containing at least 'Label', 'X', 'Y', 'Entity', 'Cluster'.}

\item{new_coordinates}{a data frame containing at least 'Label', 'newX', 'newY'.}

\item{incidence_matrix}{a data frame representing incidence matrix of the network. Need to have row names and column names. The label order of the incidence matrix needs to be the same as in nodelist.}

\item{plotlabel}{which side to keep node label. Can be 'col', 'row', 'both', 'neither'. Default is 'col'.}

\item{nodesize_min}{minimum size of a node. Also used if node size is not scaled by degree. Default is 1.}

\item{nodesize_max}{maximum size of a node. Default is 10}

\item{nodesize_by_degree}{Scale the degree by 'row', 'col', 'both', or 'neither'. Default is 'row'}

\item{nodeborder_row}{border thickness of row nodes. Default is 0.1}

\item{nodeborder_col}{border thickness of column nodes. Default is 2}

\item{nodelabelsize}{label size of nodes. Default is 3.}

\item{nodelabelcolor}{label color of nodes. Default is grey20.}
}
\value{
a data frame containing 'Label', 'X', 'Y', 'Entity', 'Cluster', 'Degree', 'Size', 'Shape', 'Border', 'Color', 'Plotlabel', 'Labelsize'.
}
\description{
Get nodelist with extra features and plotting information. It overwrites the function with same name in ExplodeLayout for more detailed control.
}
