% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{get_significance_zp}
\alias{get_significance_zp}
\title{Get significance of a real value over a Gaussian distribution.}
\usage{
get_significance_zp(V_real, V_randomdist)
}
\arguments{
\item{V_real}{real value.}

\item{V_randomdist}{a vector of values, assuming they follow a Gaussian distribution.}
}
\value{
V_real original real value

V_random_mean mean of V_randomdist

V_random_sd standard deviation of V_randomdist

z_score z-score of V_real compared to V_randomdist

p_val two-sided p-value of z-score.

V_random_length length of V_randomdist

V_random_min minimum of V_randomdist

V_random_max maximum of V_randomdist

V_random_median median of V_randomdist

V_random_mad median absolute deviation of V_randomdist
}
\description{
Get significance of a real value over a Gaussian distribution.
}
