% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cluster_cmp_OR.R
\name{pairwise_clu_compare}
\alias{pairwise_clu_compare}
\title{Cluster pairwise odds ratio for binary outcome using logistic regression, optionally control for covariates.}
\usage{
pairwise_clu_compare(
  data,
  cluster_colname,
  outcome_colname,
  covariate_colname = c()
)
}
\arguments{
\item{data}{a dataframe containing cluster, outcome, and optionally covariates.}

\item{cluster_colname}{name of a column containing clustering information. Number of unique clusters should be at least 2.}

\item{outcome_colname}{name of a column containing binary outcome.}

\item{covariate_colname}{a vector of names of columns for covariates. Default is empty.}
}
\value{
a dataframe containing enrichment information
}
\description{
Cluster pairwise odds ratio for binary outcome using logistic regression, optionally control for covariates.
}
