% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{pipeline_Mod_ModSig_EL}
\alias{pipeline_Mod_ModSig_EL}
\title{A simple pipeline finding modularity (and optionally significance of modularity) and ExplodeLayout of a bipartite network.}
\usage{
pipeline_Mod_ModSig_EL(
  fname = "example",
  EL_radius = 1.2,
  EL_nodelab = "c",
  signif_preserve = "col_deg",
  signif_N_sample = 10
)
}
\arguments{
\item{fname}{stem of file name of input and output files.}

\item{EL_radius}{radius of ExplodeLayout.}

\item{EL_nodelab}{whether to plot node labels. '' labels no nodes, 'r' labels row nodes, 'c' labels column nodes, 'rc' labels all nodes.}

\item{signif_preserve}{null models to use on incidence matrix. 'density' shuffles entries across whole matrix, 'row_deg' shuffles entries within each row, 'col_deg' shuffles entries within each column.}

\item{signif_N_sample}{number of random networks generated for significance test. The significance test is skipped if signif_N_sample<3.}
}
\value{
No return value. All results are saved in files.

This function does the following things:
1. reads bipartite network data represented as incidence matrix from file "[fname]_data.csv";
2. removes empty, a.k.a, isolated rows and columns where all entries are 0, and saves the new matrix to "[fname]_incidmat.csv";
3. finds modularity and optimal clustering of the network, and saves the nodelist with clustering membership to "[fname]_nodelist.csv";
4. optionally, finds significance of modularity using specific null model, saves the modularity distribution of the random networks generated by the null model to "[fname]_Qrandomdist.csv", and saves the significance result to "[fname]_zp_Q.csv";
5. plot the network using ExplodeLayout.
}
\description{
A simple pipeline finding modularity (and optionally significance of modularity) and ExplodeLayout of a bipartite network.
}
