% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plot.R
\name{plot_binet_ggplot2_from_nodelist_and_edgelist}
\alias{plot_binet_ggplot2_from_nodelist_and_edgelist}
\title{Plot a network base on the provided nodelist and edgelist.}
\usage{
plot_binet_ggplot2_from_nodelist_and_edgelist(
  nodelist_for_plotting,
  edgelist_for_plotting
)
}
\arguments{
\item{nodelist_for_plotting}{a data frame containing 'Label', 'X', 'Y', 'Entity', 'Cluster', 'Degree', 'Size', 'Shape', 'Border', 'Color', 'Plotlabel', 'Labelsize'.}

\item{edgelist_for_plotting}{a data frame containing 'nodesR',    'nodesC',    'values',    'x0',    'y0',    'x1',    'y1',    'weight',    'edgewidth',    'color'.}
}
\value{
a ggplot2 plot of the network
}
\description{
Plot a network base on the provided nodelist and edgelist.
}
