% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cluster_cmp_OR.R
\name{plotOddsRatio_CaseVsCtrl}
\alias{plotOddsRatio_CaseVsCtrl}
\title{Plot odds ratio of each cluster compared to control (cluster 0).}
\usage{
plotOddsRatio_CaseVsCtrl(pcc)
}
\arguments{
\item{pcc}{odds ratio dataframe generated from function 'pairwise_clu_compare' or 'enrich'.}
}
\value{
a plot of odds ratio compared to control (cluster 0) with confidence interval as error bar.
}
\description{
Plot odds ratio of each cluster compared to control (cluster 0).
}
