% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cluster_cmp_OR.R
\name{print_OddsRatio}
\alias{print_OddsRatio}
\title{Generate strings summarizing odds ratio, confidence interval, and p-value.}
\usage{
print_OddsRatio(pcc)
}
\arguments{
\item{pcc}{odds ratio dataframe generated from function 'pairwise_clu_compare'}
}
\value{
a vector of strings summarizing odds ratio, confidence interval, and p-value.
}
\description{
Generate strings summarizing odds ratio, confidence interval, and p-value.
}
