% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cluster_cmp_OR.R
\name{psig}
\alias{psig}
\title{Add text behind p-value.}
\usage{
psig(p)
}
\arguments{
\item{p}{p-value}
}
\value{
string describing whether p<.001, p<.01, or p<.05.
}
\description{
Add text behind p-value.
}
