% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{remove_isolatedNodes}
\alias{remove_isolatedNodes}
\title{Remove isolated (all zero) rows and columns of an incidence matrix.}
\usage{
remove_isolatedNodes(incid_mat)
}
\arguments{
\item{incid_mat}{incidence matrix.}
}
\value{
a new matrix with isolated rows and columns removed.
}
\description{
Remove isolated (all zero) rows and columns of an incidence matrix.
}
