% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{shuffle_incid_mat}
\alias{shuffle_incid_mat}
\title{Shuffle incidence matrix using different null models.}
\usage{
shuffle_incid_mat(incid_mat, preserve = c("density", "row_deg", "col_deg"))
}
\arguments{
\item{incid_mat}{incidence matrix}

\item{preserve}{'density' shuffles entries across whole matrix, 'row_deg' shuffles entries within each row, 'col_deg' shuffles entries within each column.}
}
\value{
shuffled incidence matrix
}
\description{
Shuffle incidence matrix using different null models.
}
