% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{simulate_modularity_distribution_for_null_model}
\alias{simulate_modularity_distribution_for_null_model}
\title{Get modularity distribution of random networks generated from an original network using specific null model.}
\usage{
simulate_modularity_distribution_for_null_model(
  incid_mat,
  N_sample = 1000,
  preserve = "col_deg"
)
}
\arguments{
\item{incid_mat}{incidence matrix of original network.}

\item{N_sample}{number of samples, default to 1000.}

\item{preserve}{'density' shuffles entries across whole matrix, 'row_deg' shuffles entries within each row, 'col_deg' shuffles entries within each column.}
}
\value{
Q a vector of modularity of the randomly generated networks.

Nclu a vector of number of clusters of the randomly generated networks.
}
\description{
Get modularity distribution of random networks generated from an original network using specific null model.
}
