% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clu_BS_stablility.R
\name{utility_BS_Stability}
\alias{utility_BS_Stability}
\title{Measure clustering stability of columns by bootstrapping rows (sample same number of rows with replacement).}
\usage{
utility_BS_Stability(
  inputName_nodelist = "example_nodelist.csv",
  inputName_incidmat = "example_incidmat.csv",
  inputName_modularity = "example_modularity.csv",
  BS_N_sample = 10,
  outputName_bs_randomdist = "example_bs_randomdist.csv",
  outputName_bs_zp = "example_bs_zp.csv"
)
}
\arguments{
\item{inputName_nodelist}{name of nodelist file}

\item{inputName_incidmat}{name of incidence matrix file}

\item{inputName_modularity}{name of a csv file saving real modularity Q}

\item{BS_N_sample}{number of bootstrap samples. Default to 10 as a proof of concept. Should be at least 1000 in practice.}

\item{outputName_bs_randomdist}{name of a csv file to save modularity, number of clusters, and Rand Index compared to the original column clustering for each bootstrapped network}

\item{outputName_bs_zp}{name of a csv file to save modularity significance result including z-score, p-value and other information.}
}
\value{
a data frame containing modularity significance result including z-score, p-value and other information.
}
\description{
Measure clustering stability of columns by bootstrapping rows (sample same number of rows with replacement).
}
