% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{utility_EL}
\alias{utility_EL}
\title{Plot network from data file using ExplodeLayout and save the plot to pdf}
\usage{
utility_EL(fname, radius, plotlabel)
}
\arguments{
\item{fname}{stem of file name of input and output files.}

\item{radius}{radius of ExplodeLayout.}

\item{plotlabel}{whether to plot node labels. '' labels no nodes, 'r' labels row nodes, 'c' labels column nodes, 'rc' labels all nodes.}
}
\value{
p the ggplot object

This function does the following things:
1. reads nodelist file from "[fname]_nodelist.csv";
2. reads incidence matrix file from "[fname]_incidmat.csv";
3. visualizes the network using ExplodeLayout according to user specified radius and prints the plot on screen;
4. saves the plot to "[fname]_EL_r[radius].pdf"
}
\description{
Plot network from data file using ExplodeLayout and save the plot to pdf
}
