% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_EL.R
\name{utility_EL_advanced}
\alias{utility_EL_advanced}
\title{Generate nodelist and edgelist for plotting using ExplodeLayout.}
\usage{
utility_EL_advanced(
  inputName_nodelist = "example_nodelist.csv",
  inputName_incidmat = "example_incidmat.csv",
  radius = 1.5,
  outputName_nodelist = "example_nodelist_plot.csv",
  outputName_edgelist = "example_edgelist_plot.csv",
  plotlabel = "col",
  nodesize_min = 1,
  nodesize_max = 10,
  nodesize_by_degree = "row",
  nodeborder_row = 0.1,
  nodeborder_col = 2,
  nodelabelsize = 3,
  nodelabelcolor = "grey20",
  edgethickness_min = 0.01,
  edgethickness_max = 1,
  edgegrayscale_min = 0.5,
  edgegrayscale_max = 1,
  edgethickness_by_weight = TRUE
)
}
\arguments{
\item{inputName_nodelist}{a csv file containing 'Label', 'X', 'Y', 'Entity', 'Cluster'. Can be generated using 'utility_Mod'.}

\item{inputName_incidmat}{a csv file representing incidence matrix of the network. Can be generated using 'utility_Mod'.}

\item{radius}{radius of ExplodeLayout. Default is 1.5}

\item{outputName_nodelist}{output name of the enhanced nodelist file for plotting.}

\item{outputName_edgelist}{output name of the enhanced edgelist file for plotting.}

\item{plotlabel}{which side to keep node label. Can be 'col', 'row', 'both', 'neither'. Default is 'col'.}

\item{nodesize_min}{minimum size of a node. Also used if node size is not scaled by degree. Default is 1.}

\item{nodesize_max}{maximum size of a node. Default is 10}

\item{nodesize_by_degree}{Scale the degree by 'row', 'col', 'both', or 'neither'. Default is 'row'}

\item{nodeborder_row}{border thickness of row nodes. Default is 0.1}

\item{nodeborder_col}{border thickness of column nodes. Default is 2}

\item{nodelabelsize}{label size of nodes. Default is 3.}

\item{nodelabelcolor}{label color of nodes. Default is grey20.}

\item{edgethickness_min}{Also used if edgethickness_by_weight==FALSE. Default is 0.01.}

\item{edgethickness_max}{Default is 1}

\item{edgegrayscale_min}{Darkest in gray scale. 0 for black, 1 for white. Default is 0.5}

\item{edgegrayscale_max}{Brightest in gray scale. Default is 1.}

\item{edgethickness_by_weight}{if TRUE, scale edge width by edge value. Default is TRUE.}
}
\value{
a list of enhanced nodelist and edgelist for plotting.
}
\description{
Generate nodelist and edgelist for plotting using ExplodeLayout.
}
