% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_EL.R
\name{utility_EL_plot}
\alias{utility_EL_plot}
\title{Plot a network using nodelist file and edgelist file, and save the plot.}
\usage{
utility_EL_plot(
  inputName_nodelist = "example_nodelist_plot.csv",
  inputName_edgelist = "example_edgelist_plot.csv",
  outputName = "example_plot",
  outputType = c("svg", "pdf", "ps", "bmp", "jpg", "png", "tif")
)
}
\arguments{
\item{inputName_nodelist}{a csv file containing 'Label', 'X', 'Y', 'Entity', 'Cluster', 'Degree', 'Size', 'Shape', 'Border', 'Color', 'Plotlabel', 'Labelsize'. Can be generated using function 'utility_EL_advanced'.}

\item{inputName_edgelist}{a csv file containing 'nodesR',    'nodesC',    'values',    'x0',    'y0',    'x1',    'y1',    'weight',    'edgewidth',    'color'. Can be generated using function 'utility_EL_advanced'.}

\item{outputName}{name of output file without extension.}

\item{outputType}{format to be saved. Currently supports 'svg', 'pdf', 'ps', 'bmp', 'jpg', 'png', 'tif'. Also used as extension of the output file name.}
}
\value{
a ggplot2 plot of the network
}
\description{
Plot a network using nodelist file and edgelist file, and save the plot.
}
