% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_Mod.R
\name{utility_Mod}
\alias{utility_Mod}
\title{Find optimal partition (clustering) and corresponding bipartite modularity of a bipartite network represented by incidence matrix.}
\usage{
utility_Mod(
  inputName_data = "example_data.csv",
  outputName_nodelist = "example_nodelist.csv",
  outputName_incidmat = "example_incidmat.csv",
  outputName_modularity = "example_modularity.csv"
)
}
\arguments{
\item{inputName_data}{name of a csv file of incidence matrix of bipartite network where the first column is row name.}

\item{outputName_nodelist}{name of a csv file to save nodelist.}

\item{outputName_incidmat}{name of a csv file to save incidence matrix after removing isolated rows and columns.}

\item{outputName_modularity}{name of a csv file to save modularity Q.}
}
\value{
MODULARITY modulairty

ASSIGN partition of the network after removing isolated nodes.


This function does the following things:
1. reads bipartite network data represented as incidence matrix from [inputName_data];
2. removes empty, a.k.a, isolated rows and columns where all entries are 0, and saves the new matrix to [outputName_incidmat];
3. finds modularity and optimal clustering of the network, and saves the nodelist with clustering membership to [outputName_nodelist];
4. save modularity to [outputName_modularity];
5. return modularity and partition of the network after removing isolated nodes.
}
\description{
Find optimal partition (clustering) and corresponding bipartite modularity of a bipartite network represented by incidence matrix.
}
