% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{utility_Mod_ModSig}
\alias{utility_Mod_ModSig}
\title{A simple pipeline finding optimal clustering and corresbonding modularity (and optionally significance of modularity).}
\usage{
utility_Mod_ModSig(
  fname = "example",
  signif_preserve = "col_deg",
  signif_N_sample = 10
)
}
\arguments{
\item{fname}{stem of file name of input and output files.}

\item{signif_preserve}{null models to use on incidence matrix. 'density' shuffles entries across whole matrix, 'row_deg' shuffles entries within each row, 'col_deg' shuffles entries within each column.}

\item{signif_N_sample}{number of random networks generated for significance test. The significance test is skipped if signif_N_sample<3.}
}
\value{
No return value. All results are saved in files.

This function does the following things:
1. reads bipartite network data represented as incidence matrix from file "[fname]_data.csv";
2. removes empty, a.k.a, isolated rows and columns where all entries are 0, and saves the new matrix to "[fname]_incidmat.csv";
3. finds modularity and optimal clustering of the network, and saves the nodelist with clustering membership to "[fname]_nodelist.csv";
4. optionally, finds significance of modularity using specific null model, saves the modularity distribution of the random networks generated by the null model to "[fname]_Qrandomdist.csv", and saves the significance result to "[fname]_zp_Q.csv";
}
\description{
A simple pipeline finding optimal clustering and corresbonding modularity (and optionally significance of modularity).
}
