% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{vector_CenterScale}
\alias{vector_CenterScale}
\title{Center-Scale transformation of a vector}
\usage{
vector_CenterScale(x)
}
\arguments{
\item{x}{a numeric vector of length at least 2 without missing values and are not identical.}
}
\value{
Center-Scale transformed vector as (x-mean(x))/sd(x)
}
\description{
Center-Scale transformation of a vector
}
