% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_EL.R
\name{utility_EL}
\alias{utility_EL}
\title{Plot network from data file using ExplodeLayout and save the plot to pdf}
\usage{
utility_EL(
  inputName_nodelist = "example_nodelist.csv",
  inputName_incidmat = "example_incidmat.csv",
  radius = 1.5,
  plotlabel = "c",
  outputName_plotpdf = "example_plot.pdf"
)
}
\arguments{
\item{inputName_nodelist}{name of nodelist file}

\item{inputName_incidmat}{name of incidence matrix file}

\item{radius}{radius of ExplodeLayout.}

\item{plotlabel}{whether to plot node labels. '' labels no nodes, 'r' labels row nodes, 'c' labels column nodes, 'rc' labels all nodes.}

\item{outputName_plotpdf}{name of the pdf to plot.}
}
\value{
the ggplot object


This function does the following things:
1. reads nodelist file from [inputName_nodelist];
2. reads incidence matrix file from [inputName_incidmat];
3. visualizes the network using ExplodeLayout according to user specified radius and prints the plot on screen;
4. saves the plot to [outputName_plotpdf].
}
\description{
Plot network from data file using ExplodeLayout and save the plot to pdf
}
