% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesDIVA.R
\name{vector_MinMax}
\alias{vector_MinMax}
\title{MinMax transformation of a vector}
\usage{
vector_MinMax(x)
}
\arguments{
\item{x}{a numeric vector of length at least 2 without missing values}
}
\value{
MinMax transformed vector as (x-min(x))/(max(x)-min(x))
}
\description{
MinMax transformation of a vector
}
